move batch iterator to separate file
