+++ lib/commands/unipept.rb
@@ -3,6 +3,7 @@ require 'typhoeus'
 require_relative '../formatters'
 require_relative '../configuration'
 require_relative '../batch_order'
+require_relative '../batch_iterator'
 require_relative '../version'
 
 require_relative 'unipept/config'
